+++ src/commonMain/kotlin/csense/kotlin/extensions/Exceptions.kt
@@ -3,6 +3,7 @@
 package csense.kotlin.extensions
 
 import csense.kotlin.*
+import csense.kotlin.extensions.generic.*
 import csense.kotlin.logger.*