+++ src/commonMain/kotlin/csense/kotlin/logger/PrintLoggers.kt
@@ -1,0 +1,18 @@
+package csense.kotlin.logger
+
+import kotlinx.coroutines.*
+
+
+@OptIn(DelicateCoroutinesApi::class)
+public fun CsenseLogger.printLogsToConsole(scope: CoroutineScope = GlobalScope): Job = scope.launch {
+    allLoggers.collect {
+        println(it.toFullLogString())
+    }
+}
+
+@OptIn(DelicateCoroutinesApi::class)
+public fun CsenseLogger.printLogsToConsoleAnsiColored(scope: CoroutineScope = GlobalScope): Job = scope.launch {
+    allLoggers.collect {
+        println(it.toFullColoredLog())
+    }
+}