+++ src/jvmMain/kotlin/csense/kotlin/extensions/JvmThrowable.kt
@@ -2,6 +2,7 @@
 
 package csense.kotlin.extensions
 
+import csense.kotlin.extensions.generic.*
 import java.io.*
 
 /**
@@ -18,4 +19,40 @@ public inline fun Throwable.stackTraceToString(): String {
         }
     }
     return writer.toString()
+}
+
+
+public inline fun Throwable.toSensitiveStackTraceString(): String {
+    return formatToString(
+        onException = { "${it::class.java.canonicalName} -message excluded-" },
+        onStackTrace = {
+            "\t at $it"
+        },
+        onSuppressed = {
+            "Suppressed - ${it::class.java.canonicalName}"
+        }
+    )
+}
+
+
+public inline fun Throwable.formatToString(
+    onException: (Throwable) -> String,
+    onStackTrace: (StackTraceElement) -> String,
+    onSuppressed: (Throwable) -> String,
+): String {
+    val result = StringBuilder()
+    Generic.traverseWhileNotNullAndNoCycles(
+        start = this,
+        processCurrentLevel = {
+            result.appendLineIfNotBlank(onException(it))
+            it.stackTrace.forEach { stackTraceElement ->
+                result.appendLineIfNotBlank(onStackTrace(stackTraceElement))
+            }
+            it.suppressed.forEach { suppressedElement ->
+                result.appendLineIfNotBlank(onSuppressed(suppressedElement))
+            }
+        },
+        getNextLevel = { it.cause }
+    )
+    return result.toString()
 }